Caring/requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.0
djangorestframework==3.11.1
idna==2.10
pytz==2020.1
requests==2.24.0
urllib3==1.25.10
